refactor(hosted-components): extract onboarding page and auth barrel exports
